Move TelemetryEvents class in self-contained systems to Features namespace
