+++ application/account-management/Core/Features/Authentication/Commands/RefreshAuthenticationTokens.cs
@@ -4,7 +4,6 @@ using JetBrains.Annotations;
 using Mapster;
 using Microsoft.AspNetCore.Http;
 using PlatformPlatform.AccountManagement.Features.Users.Domain;
-using PlatformPlatform.AccountManagement.TelemetryEvents;
 using PlatformPlatform.SharedKernel.Authentication;
 using PlatformPlatform.SharedKernel.Authentication.TokenGeneration;
 using PlatformPlatform.SharedKernel.Cqrs;
