+++ application/account-management/Core/Features/Users/Commands/DeleteUser.cs
@@ -1,6 +1,5 @@
 using JetBrains.Annotations;
 using PlatformPlatform.AccountManagement.Features.Users.Domain;
-using PlatformPlatform.AccountManagement.TelemetryEvents;
 using PlatformPlatform.SharedKernel.Cqrs;
 using PlatformPlatform.SharedKernel.Domain;
 using PlatformPlatform.SharedKernel.Telemetry;
