+++ application/back-office/Core/Features/TelemetryEvents.cs
